+++ lib/features/support/resilient_driver.rb
@@ -7,26 +7,13 @@ require_relative './logger'
 #
 # For methods available on this class, @see AppAutomateDriver.
 class ResilientAppiumDriver
-  # Creates the ResilientAppiumDriver
+  # Creates the AppiumDriver
   #
-  # @param username [String] the BrowserStack username
-  # @param access_key [String] the BrowserStack access key
-  # @param local_id [String] the identifier for the BrowserStackLocal tunnel
-  # @param target_device [String] a key from the Devices array selecting which device capabilities to target
-  # @param app_location [String] the location of the test-app to upload
-  # @param locator [Symbol] the primary locator strategy Appium should use to find elements
-  # @param additional_capabilities [Hash] a hash of additional capabilities to be used in this test run
-  def initialize(username, access_key, local_id, target_device, app_location, locator = :id, additional_capabilities = {})
-    @driver = AppAutomateDriver.new username,
-                                    access_key,
-                                    local_id,
-                                    target_device,
-                                    app_location,
-                                    locator,
-                                    additional_capabilities
-
-    # This must appear after creation of @driver as AppAutomateDriver also does this
-    MazeRunner.driver = self
+  # @param server_url [String] URL of the Appium server
+  # @param capabilities [Hash] a hash of capabilities to be used in this test run
+  def initialize(server_url, capabilities)
+    @driver = AppiumDriver.new server_url,
+                               capabilities
   end
 
   def respond_to_missing?(method_name, include_private = false)
@@ -42,7 +29,9 @@ class ResilientAppiumDriver
         return @driver.send(method, *args, &block)
       rescue Selenium::WebDriver::Error::UnknownError, Selenium::WebDriver::Error::WebDriverError => error
         retries += 1
+        $logger.warn error
         $logger.warn 'Appium Error occurred - restarting driver.'
+        sleep 3
         restart
       end
     end